feat: add ability to delete product, variant
when deleting a product or a variant,
soft deletion is performed,
i.e. only the isActive field is changed to false.
This is due to the fact that
it will be problematic in the future to keep track of products
that have been added to the cart or are already in the order.
Images are not deleted at all, neither from the database nor from Cloudinary.
Deleting a product implies deleting all variants that can be restored in the future.