feat: add decorator to pull organizationId from headers
now organizationId is sent to headers from the frontend
and the response to the request to get all stores sends only those
that have the same organizationId as the one the user has.

It's also worth noting that if undefined data is sent to "where",
Prisma treats it as missing,
so I had to put an empty string if suddenly the header didn't come up in the query.
See: https://www.prisma.io/docs/concepts/components/prisma-client/null-and-undefined